+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "test/erpc_module"]
-	path = test/erpc_module
+	path = std-transport/test/erpc_module
 	url = https://github.com/EmbeddedRPC/erpc
