+++ src/lib.rs
@@ -1,6 +1,5 @@
-#![no_std]
-#[cfg(feature = "std")]
-extern crate std;
+
+#![cfg_attr(not(test), no_std)]
 
 pub mod codec;
 pub mod cursor;
