+++ std-transport/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "std-transport"
+version = "0.1.0"
+authors = ["tomoyuki-nakabayashi <tomo.wait.for.it.yuki@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rust-erpc = {path = "../"}
+serialport = "3.3.0"
