+++ std-transport/test/erpc_shim/rpc_ble_api/common.py
