+++ std-transport/test/erpc_shim/rpc_ble_callback/common.py
